refactor: selector 파일 분리
